Split interop/tsgo.rs probe generation into phases
Splits tsgo.rs from 2166 to 531 lines by extracting 4 sub-modules:

- probe_gen.rs (1204 lines): generate_probe + all helpers
- probe_run.rs (150 lines): create_probe_dir, run_tsgo, find_probe_dts
- specifier_map.rs (215 lines): build_specifier_map
- typeof_resolve.rs (109 lines): resolve_typeof_types, find_package_dts

Uses Rust 2018+ module style. Removed duplicated helper functions
(unwrap_try_await_expr, expr_to_callee_name) by making them pub(super)
in probe_gen and importing in specifier_map.

Purely structural - no behavioral changes.
